[ALPS03595399] VPU: fix 32b ioctrl issue
[Detail] fix 32bit ioctrl issue

[Solution] fix 32bit ioctrl issue

MTK-Commit-Id: 4e58c8af9b62303d4df9e3ad7b5304efef5ea49d

CR-Id: ALPS03595399
Feature: MT6771 Platform Feature

+++ drivers/misc/mediatek/vpu/2.0/vpu_drv.c
@@ -630,7 +630,7 @@ static long vpu_ioctl(struct file *flip, unsigned int cmd, unsigned long arg)
 
 		u_req = (struct vpu_request *) arg;
 
-		ret = copy_from_user(&req->request_id, &u_req->request_id, sizeof(uint64_t));
+		ret = copy_from_user(&req->request_id, &u_req->request_id, sizeof(uintptr_t));
 		CHECK_RET("[REG] copy 'req id' failed,%d\n", ret);
 
 		LOG_DBG("[vpu] deque test: user_id_0x%lx, request_id_0x%lx", (unsigned long)user,
